Avoid calling unlink on directories
Be sure to not call unlink on directories, because such operation may
succeed in some systems (such as SunOS with UFS) when running as root
and we do not expect that behavior.

Monotone-Parent: 6cee9fb5f59cff9596ee7673bddf7fed2dbd009c
Monotone-Revision: 23bb42808b08f16a4eb0c76c08223dfaf2030119

Monotone-Author: jmmv@NetBSD.org
Monotone-Date: 2008-02-02T21:40:13
Monotone-Branch: org.NetBSD.atf.src

+++ atf/fs.cpp
@@ -561,6 +561,10 @@ impl::is_executable(const path& p)
 void
 impl::remove(const path& p)
 {
+    if (file_info(p).get_type() == file_info::dir_type)
+        throw atf::system_error(IMPL_NAME "::remove(" + p.str() + ")",
+                                "Is a directory",
+                                EPERM);
     if (::unlink(p.c_str()) == -1)
         throw atf::system_error(IMPL_NAME "::remove(" + p.str() + ")",
                                 "unlink(" + p.str() + ") failed",
